+++ shared/modules/persistence_module.go
@@ -76,7 +76,7 @@ type PersistenceWriteContext interface {
 
 	// Indexes the transaction using several different keys (for lookup purposes) in the key-value store
 	// that backs the transaction merkle tree.
-	IndexTransaction(txResult *coreTypes.TxResult) error
+	IndexTransaction(idxTx *coreTypes.IndexedTransaction) error
 
 	// Pool Operations
 	AddPoolAmount(address []byte, amount string) error