+++ consensus/doc/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.50] - 2023-05-03
+
+- Rename `TxResult` to `IndexedTransaction`
+
 ## [0.0.0.49] - 2023-04-27
 
 - Updated comments following P2P refactor
@@ -17,7 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.0.47] - 2023-04-17
 
--  Log warnings in `handleStateSyncMessage()` 
+-  Log warnings in `handleStateSyncMessage()`
 
 ## [0.0.0.46] - 2023-04-13